flask_backend/requirements.txt
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astroid==2.3.3
bleach==2.1.2
bootstrap4==0.1.0
Brlapi==0.6.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2
django-allauth==0.41.0
django-bootstrap-modal-forms==1.5.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-jsonfield==1.4.0
django-widget-tweaks==1.4.8
entrypoints==0.2.3.post1
fail2ban==0.10.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
html5lib==0.999999999
httplib2==0.9.2
idna==2.10
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.3
mysqlclient==1.4.6
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
notebook==5.2.2
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
pika==1.1.0
Pillow==7.1.1
prompt-toolkit==1.0.15
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.14.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
ssh-import-id==5.7
stripe==2.48.0
system-service==0.3
systemd-python==234
terminado==0.7
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==1.0.1
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
